fix(components): add a confirmation dialog for deleting team lib photo (#4925)

+++ app/views/components/ImageSelectDialog/Tabs/TeamLibrary/index.tsx
@@ -98,7 +98,17 @@ function TeamLibrary({ query, onSelect, onEdit }: SearchableImageTabProps) {
   }
 
   const handleDelete = (asset: IBrandAsset) => {
-    deleteAsset(asset.id)
+    dispatch(
+      confirmation({
+        message: 'Delete photo',
+        description: 'Are your sure about deleting this photo?',
+        confirmLabel: 'Yes, I do',
+        appearance: 'danger',
+        onConfirm: () => {
+          deleteAsset(asset.id)
+        }
+      })
+    )
   }
 
   if (isLoading) {
